requirements.txt
Django==2.0.5
djangorestframework_jwt==1.10.0
djangorestframework==3.8.2
pydash==4.7.4
django-url-filter==0.3.12